+++ src/main/generate.cljs
@@ -1,16 +1,18 @@
 (ns main.generate
   (:require
     [cljs.nodejs :as node :refer [enable-util-print!]]
-    [cljstron.browser.app :refer [pprint-str def-shadow-cljs-edn]]
     [cljs.pprint :refer [pprint]]
-    [cljstron.browser.app :refer [read-json read-app-edn]]))
+    [cljstron.browser.app :refer [write-gen-files read-json read-app-edn]]))
 
 (enable-util-print!)
 
 (defn main []
   "Generate compile files for the project"
-  (pprint (read-app-edn :main "app.edn"))
+  (pprint (read-app-edn :main "."))
   (println "-----------------------------------")
   (pprint (read-json "package.json"))
   (println "-----------------------------------")
-  (pprint (def-shadow-cljs-edn)))
+  (pprint (read-app-edn :-gen-files "."))
+  (println "-----------------------------------")
+  (write-gen-files ".")
+  (println "generated files"))